fix: use postinstall hook to patch anidb-udp-client
There are a few bugs in anidb-udp-client, some can be worked around in code some cannot.

We (ab)use the `postinstall` to use sed to patch one of the more annoying bugs that we cannot work around another way.
Additionall we also trigger a compile so we generate a nice static binary.

Lists of open PRs:
- https://github.com/tsukeero/anidb-udp-client/pull/21
- https://github.com/tsukeero/anidb-udp-client/pull/22
- https://github.com/tsukeero/anidb-udp-client/pull/23